requirements.txt
asgiref==3.5.2
Django==4.1.4
django-autoslug==1.9.8
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
